+++ http4k-format/moshi/src/main/kotlin/org/http4k/format/deprecatedFormatCore.kt
@@ -1,0 +1,4 @@
+package org.http4k.format
+
+@Deprecated("Renamed")
+typealias CollectionEdgeCasesAdapter = MapAdapter