+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
-### v4.17.6.0 (uncut)
+### v4.17.6.0
+- **http4k-format-moshi-yaml*** : New module! YAML marshalling with zero-reflection is now possible due to a combination of Moshi and SnakeYaml
 - **http4k-format-jackson-xml** : Fix to HttpEvent to use correct value in xUriTemplate instead of full path.
 add autoBody for ConfigurableJacksonXml. H/T @oharaandrew314
 
@@ -1787,7 +1788,7 @@ to `RequestContextKey.of()` with `RequestContextKey.required()`
 - Fix for #24 - UriTemplate captures query parameters when the trailing path parameter is a regex.
 
 ### v2.1.0
-- Added GSON full-auto functions to convert arbitary objects to/from JSON.
+- Added GSON full-auto functions to convert arbitrary objects to/from JSON.
 
 ### v2.0.5
 - Fix #23. Contract now supports multi-part URL params (for hardcoded parts)